+++ apps/backend/src/main/kotlin/ca/uvictrades/trades/controller/stock/responses/GetWalletBalanceResponse.kt
@@ -7,6 +7,6 @@ data class GetWalletBalanceResponse(
 	val data: Data,
 ) {
 	data class Data(
-		val balance: BigDecimal,
+		val balance: Int,
 	)
 }
